+++ UserPermissions.js
@@ -1,6 +1,5 @@
-// We have to remove node_modules/react to avoid having multiple copies loaded.
-// eslint-disable-next-line import/no-unresolved
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import RenderPermissions from './lib/RenderPermissions';
 
 const propTypes = {