+++ index.js
@@ -1,12 +1,11 @@
-// We have to remove node_modules/react to avoid having multiple copies loaded.
-// eslint-disable-next-line import/no-unresolved
-import React, { Component, PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import Route from 'react-router-dom/Route';
 import Switch from 'react-router-dom/Switch';
 import Users from './Users';
 import Settings from './settings';
 
-class UsersRouting extends Component {
+class UsersRouting extends React.Component {
   static actionNames = ['stripesHome', 'usersSortByName'];
 
   static propTypes = {