+++ settings/permissions/PermissionSet.js
@@ -1,8 +1,6 @@
-// We have to remove node_modules/react to avoid having multiple copies loaded.
-// eslint-disable-next-line import/no-unresolved
-import React, { PropTypes } from 'react';
-import { connect } from '@folio/stripes-connect'; // eslint-disable-line
-import RenderPermissions from '../lib/RenderPermissions';
+import React from 'react';
+import PropTypes from 'prop-types';
+import RenderPermissions from '../../lib/RenderPermissions';
 
 class PermissionSet extends React.Component {
   static propTypes = {