+++ settings/permissions/PermissionSets.js
@@ -1,6 +1,5 @@
-// We have to remove node_modules/react to avoid having multiple copies loaded.
-// eslint-disable-next-line import/no-unresolved
-import React, { Component, PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 
 import _ from 'lodash';
 
@@ -13,7 +12,7 @@ import NavListSection from '@folio/stripes-components/lib/NavListSection';
 import IfPermission from '@folio/stripes-components/lib/IfPermission';
 import PermissionSetDetails from './PermissionSetDetails';
 
-class PermissionSets extends Component {
+class PermissionSets extends React.Component {
   static propTypes = {
     stripes: PropTypes.shape({
       hasPerm: PropTypes.func.isRequired,
@@ -22,10 +21,9 @@ class PermissionSets extends Component {
       permissionSets: PropTypes.arrayOf(PropTypes.object),
     }).isRequired,
     mutator: PropTypes.shape({
-      permissionSets: PropTypes.shape({
+      updater: PropTypes.shape({
         POST: PropTypes.func,
         DELETE: PropTypes.func,
-        GET: PropTypes.func,
       }),
     }).isRequired,
   };
@@ -34,15 +32,11 @@ class PermissionSets extends Component {
     permissionSets: {
       type: 'okapi',
       records: 'permissions',
-      DELETE: {
-        path: 'perms/permissions',
-      },
-      POST: {
-        path: 'perms/permissions',
-      },
-      GET: {
-        path: 'perms/permissions?length=100&query=(mutable=true)',
-      },
+      path: 'perms/permissions?length=1000&query=(mutable=true)&expandSubs=true',
+    },
+    updater: {
+      type: 'okapi',
+      records: 'permissions',
       path: 'perms/permissions',
     },
   });
@@ -50,10 +44,9 @@ class PermissionSets extends Component {
   constructor(props) {
     super(props);
 
-    // 'Manager' just for example...
-
     this.state = {
       selectedSet: null,
+      newSet: false,
     };
 
     this.navList = null;
@@ -61,6 +54,7 @@ class PermissionSets extends Component {
     this.onSelectSet = this.onSelectSet.bind(this);
     this.createNewPermissionSet = this.createNewPermissionSet.bind(this);
     this.clearSelection = this.clearSelection.bind(this);
+    this.recordHasBeenCreated = this.recordHasBeenCreated.bind(this);
   }
 
   componentDidUpdate(prevProps) {
@@ -75,9 +69,7 @@ class PermissionSets extends Component {
 
       // Jeremy has investigated that and confirmed that it is harmless.
       // eslint-disable-next-line react/no-did-update-set-state
-      this.setState({
-        selectedSet: newPermSet,
-      });
+      this.setSelectedSet(newPermSet);
     }
   }
 
@@ -101,10 +93,13 @@ class PermissionSets extends Component {
     this.setSelectedSet(null);
   }
 
+  // It's ugly that we need this. Too much state shared between this and <PermissionSetDetails>
+  recordHasBeenCreated() {
+    this.setState({ newSet: false });
+  }
+
   createNewPermissionSet() {
-    this.props.mutator.permissionSets.POST({
-      mutable: true,
-    });
+    this.setState({ newSet: true });
   }
 
   render() {
@@ -134,7 +129,8 @@ class PermissionSets extends Component {
             </NavListSection>
           </NavList>
         </Pane>
-        {this.state.selectedSet && <PermissionSetDetails parentMutator={this.props.mutator} clearSelection={this.clearSelection} stripes={this.props.stripes} initialValues={this.state.selectedSet} selectedSet={this.state.selectedSet} />}
+        {this.state.newSet && <PermissionSetDetails parentMutator={this.props.mutator} clearSelection={this.clearSelection} stripes={this.props.stripes} selectedSet={{}} initialValues={{}} tellParentTheRecordHasBeenCreated={this.recordHasBeenCreated} />}
+        {this.state.selectedSet && !this.state.newSet && <PermissionSetDetails parentMutator={this.props.mutator} clearSelection={this.clearSelection} stripes={this.props.stripes} initialValues={this.state.selectedSet} selectedSet={this.state.selectedSet} />}
       </Paneset>
     );
   }